# SAS7BDAT parser: Improve subheader lookup performance